+++ icu4c/source/test/cintltst/ucnvseltst.c
@@ -240,7 +240,7 @@ text_open(TestText *tt) {
     fclose(f);
     return FALSE;
   }
-  if (length != fread(tt->text, 1, length, f)) {
+  if (length != (int32_t)fread(tt->text, 1, length, f)) {
     log_err("error reading %ld bytes from test text file\n", (long)length);
     length = 0;
     uprv_free(tt->text);
